requirements.txt
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
selenium
beautifulsoup4
